% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utilities.R
\name{is_grouped}
\alias{is_grouped}
\title{Check if MCC object is from grouped analysis}
\usage{
is_grouped(x)
}
\arguments{
\item{x}{An MCC object}
}
\value{
Logical indicating whether the analysis was grouped
}
\description{
Check if MCC object is from grouped analysis
}
