% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{is_mcc}
\alias{is_mcc}
\title{Check if object is an \verb{<mcc>} result}
\usage{
is_mcc(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
TRUE if x is an MCC object, FALSE otherwise
}
\description{
Check if object is an \verb{<mcc>} result
}
