% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{mcc_estimates}
\alias{mcc_estimates}
\title{Extract MCC estimates from \verb{<mcc>} objects}
\usage{
mcc_estimates(x, ...)
}
\arguments{
\item{x}{An MCC object}

\item{...}{Additional arguments (currently unused)}
}
\value{
A tibble with MCC estimates
}
\description{
Extract MCC estimates from \verb{<mcc>} objects
}
