% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utilities.R
\name{mcc_grouping_var}
\alias{mcc_grouping_var}
\title{Get grouping variable name from grouped MCC object}
\usage{
mcc_grouping_var(x)
}
\arguments{
\item{x}{An MCC object}
}
\value{
Character string with grouping variable name, or NULL if not grouped
}
\description{
Get grouping variable name from grouped MCC object
}
