% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utilities.R
\name{mcc_method}
\alias{mcc_method}
\title{Get the method used for MCC calculation}
\usage{
mcc_method(x)
}
\arguments{
\item{x}{An MCC object}
}
\value{
Character string indicating the method ("equation" or "sci")
}
\description{
Get the method used for MCC calculation
}
