% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccount-package.R
\docType{package}
\name{mccount-package}
\alias{mccount}
\alias{mccount-package}
\title{mccount: Estimate Recurrent Event Burden with Competing Risks}
\description{
Provides tools for estimating the Mean Cumulative Count (MCC), a statistical method for quantifying the burden of recurrent events in the presence of competing risks. The package implements two complementary estimation approaches: the direct probability-based 'equation' method and the 'sum of cumulative incidence' method. Both approaches handle right censoring, with additional support for left truncation in the latter. Features include automatic adjustment for simultaneous events, detailed calculation tables, and utilities for bootstrapping confidence intervals. Methods are implemented following Dong et al. (2015) \doi{10.1093/aje/kwu289}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KennethATaylor/mccount}
  \item \url{https://kennethataylor.github.io/mccount/}
  \item Report bugs at \url{https://github.com/KennethATaylor/mccount/issues}
}

}
\author{
\strong{Maintainer}: Kenneth A. Taylor \email{kenneth.taylor.dpt@gmail.com} (\href{https://orcid.org/0000-0002-3205-9280}{ORCID})

}
\keyword{internal}
