% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{print.mcc}
\alias{print.mcc}
\title{Print method for MCC objects}
\usage{
\method{print}{mcc}(x, ...)
}
\arguments{
\item{x}{An MCC object}

\item{...}{Additional arguments (currently unused)}
}
\value{
x invisibly
}
\description{
Print method for MCC objects
}
