% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utilities.R
\name{print.mcc_comparison}
\alias{print.mcc_comparison}
\title{Print method for MCC comparison objects}
\usage{
\method{print}{mcc_comparison}(x, ...)
}
\arguments{
\item{x}{An mcc_comparison object}

\item{...}{Additional arguments (currently unused)}
}
\value{
x invisibly
}
\description{
Print method for MCC comparison objects
}
