% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{summary.mcc}
\alias{summary.mcc}
\title{Summary method for \verb{<mcc>} objects}
\usage{
\method{summary}{mcc}(object, ...)
}
\arguments{
\item{object}{An MCC object}

\item{...}{Additional arguments (currently unused)}
}
\value{
A summary object with class "summary.mcc"
}
\description{
Summary method for \verb{<mcc>} objects
}
