% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{as_mcc}
\alias{as_mcc}
\title{Convert objects to \verb{<mcc>} class}
\usage{
as_mcc(x, method, weighted = FALSE, by_group = NULL, call = NULL, ...)
}
\arguments{
\item{x}{Object to convert to MCC}

\item{method}{Method used for calculation ("equation" or "sci")}

\item{weighted}{Logical indicating if weighted estimation was used}

\item{by_group}{Optional name of grouping variable}

\item{call}{Optional function call to store}

\item{...}{Additional arguments (currently unused)}
}
\value{
An MCC S3 object
}
\description{
Converts objects to MCC class. This is useful when you have calculation
results from other sources that you want to treat as MCC objects.
}
