% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.mcc}
\alias{autoplot.mcc}
\title{Auto-plot method for MCC objects}
\usage{
autoplot.mcc(x, ...)
}
\arguments{
\item{x}{An MCC object}

\item{...}{Additional arguments passed to plot.mcc}
}
\value{
A ggplot2 object
}
\description{
Convenience function that automatically creates an appropriate plot
for MCC objects. This is called when using the base R \code{plot()} function.
}
