% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utilities.R
\name{is_weighted}
\alias{is_weighted}
\title{Check if MCC object uses weighted estimation}
\usage{
is_weighted(x)
}
\arguments{
\item{x}{An MCC object}
}
\value{
Logical indicating whether weighted estimation was used
}
\description{
Check if MCC object uses weighted estimation
}
